UT refactoring
