+++ src/core/changelog/templates/changelog-template.ts
@@ -94,6 +94,7 @@ export const changelogTemplate = `
 {{#each this.modifications}}
 - {{this}}
 {{/each}}
+
 {{/each}}
 {{/if}}
 `.trim();
